aws/deploy-requirements.txt
aws-cdk.core==1.100.0
aws-cdk.aws_iam==1.100.0
aws-cdk.aws_s3==1.100.0
aws-cdk.aws_ecs==1.100.0
aws-cdk.aws_ecr==1.100.0
aws-cdk.aws_ec2==1.100.0
aws-cdk.aws_batch==1.100.0
aws-cdk.aws_stepfunctions==1.100.0
aws-cdk.aws_stepfunctions_tasks==1.100.0
boto3==1.15.5